+++ src/index.html
@@ -30,7 +30,7 @@
 
  <span>¡Paga tus cuentas de forma fácil y segura!</span>
 
-  <button onclick="seleccionarRegistro()" class="btn btnPrimary" id="seleccionarRegistro"> Registrarse </button>
+  <button  class="btn btnPrimary" id="seleccionarRegistro"> Registrarse </button>
 
   </section>
 
@@ -63,10 +63,10 @@
        
 
 
-<button onclick="irFormularioTarjeta()"  class="btn btnPrimary" id="irFormularioTarjeta"> Aceptar </button>
+<button class="btn btnPrimary" id="irFormularioTarjeta"> Aceptar </button>
 
 
-<button onclick="cancelFormularioTarjeta()" class="btnCancel" id="cancelFormularioTarjeta"> Cancelar </button>
+<button class="btnCancel" id="cancelFormularioTarjeta"> Cancelar </button>
 
 </div>
     </div>
@@ -144,8 +144,8 @@
         <input type="text" id="inputCCV" class="ccv" maxlength="3">
       </div>
     </div>
-    <button onclick="testvalidation()" class="btn btnPrimary">Validar</button>
-    <button onclick="testcancel()" class="btn btn btnCancel">Cancelar</button>
+    <button class="btn btnPrimary" id="validarTarjeta">Validar</button>
+    <button class="btn btn btnCancel" id="cancelarValidacion">Cancelar</button>
 
   </div>
 
@@ -155,7 +155,7 @@
 
    
     
-	<script src="index.js"></script>
+<script src="index.js" type="module"></script>
 
 
   </body>
